% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolators.R
\name{evalInterpolator}
\alias{evalInterpolator}
\title{Interpolator evaluation}
\usage{
evalInterpolator(ipr, x, derivative = 0)
}
\arguments{
\item{ipr}{an interpolator}

\item{x}{numeric vector giving the values to be interpolated; missing
values are not allowed}
}
\value{
Numeric vector of interpolated values.
}
\description{
Evaluation of an interpolator at some given values.
}
