% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolators.R
\name{evalInterpolator}
\alias{evalInterpolator}
\title{Interpolator evaluation}
\usage{
evalInterpolator(ipr, x, derivative = 0)
}
\arguments{
\item{ipr}{an interpolator}

\item{x}{numeric vector giving the values to be interpolated; missing
values are not allowed; for Catmull-Rom splines, the values must be
between 0 and 1}

\item{derivative}{order of differentiation, up to 1 for XXX}
}
\value{
Numeric vector of interpolated values or numeric matrix of
  interpolated points for the Catmull-Rom interpolator.
}
\description{
Evaluation of an interpolator at some given values.
}
