% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolators.R
\name{iprCatmullRom}
\alias{iprCatmullRom}
\title{Title}
\usage{
iprCatmullRom(points, closed = FALSE, alpha = 0.5)
}
\arguments{
\item{points}{matrix of points, one point per row}

\item{closed}{Boolean}

\item{alpha}{parameter between 0 and 1}
}
\value{
x
}
\description{
Title
}
\examples{
#
}
